some ground for nuget support in REPL
